set Customer entity to be mapped db tables / change id int to long

+++ src/main/java/com/havefunwith/customer/CustomerController.java
@@ -23,7 +23,7 @@ public class CustomerController {
     }
 
     @GetMapping("/api/v1/customers/{customerId}")
-    public Customer getCustomer(@PathVariable("customerId") Integer customerId) {
+    public Customer getCustomer(@PathVariable("customerId") Long customerId) {
         return customerService.getCustomer(customerId);
     }
